Add flag to reset database

+++ song_populator.py
@@ -3,30 +3,29 @@ from app.models import Song
 import os	
 import psycopg2
 
-if 'HEROKU' in os.environ:
-    DATABASE_URL = os.environ['DATABASE_URL']
-    connection = psycopg2.connect(DATABASE_URL, sslmode='require')
+if app.config['IS_HEROKU']:
+    connection = psycopg2.connect(app.config['DATABASE_URL'], sslmode='require')
 
-songs = Song.query.all()
+if app.config['RESET_DATABASE']:
+	songs = Song.query.all()
+	for s in songs:
+		db.session.delete(s)
+	db.session.commit()
 
-for s in songs:
-	db.session.delete(s)
-db.session.commit()
+	file_path = "./app/static/samplesongs.txt"
+	title = ""
+	lyrics = ""
+	with open(file_path, 'r', encoding='utf-8') as in_file:
+		for line in in_file:
+			if line.startswith("/"):
+				if lyrics != "":
+					s = Song(title=title, lyrics=lyrics)
+					db.session.add(s)
+					lyrics = ""
+				title = line.strip()[1:]
+			else:
+				lyrics += line
 
-file_path = "./app/static/samplesongs.txt"
-title = ""
-lyrics = ""
-with open(file_path, 'r', encoding='utf-8') as in_file:
-	for line in in_file:
-		if line.startswith("/"):
-			if lyrics != "":
-				s = Song(title=title, lyrics=lyrics)
-				db.session.add(s)
-				lyrics = ""
-			title = line.strip()[1:]
-		else:
-			lyrics += line
-
-s = Song(title=title, lyrics=lyrics)
-db.session.add(s)
-db.session.commit()
+	s = Song(title=title, lyrics=lyrics)
+	db.session.add(s)
+	db.session.commit()
